Reorganize e2e tests folders and files
