+++ packages/e2e-tests/specs/editor/blocks/navigation.test.js
