+++ packages/e2e-tests/specs/editor/fixtures/menu-items-response-fixture.json
