+++ packages/e2e-tests/specs/editor/various/core-settings.test.js
