+++ packages/e2e-tests/specs/editor/various/post-editor-template-mode.test.js
