+++ packages/e2e-tests/specs/experiments/navigation-editor.test.js
@@ -22,7 +22,7 @@ import { addQueryArgs } from '@wordpress/url';
 /**
  * Internal dependencies
  */
-import { useExperimentalFeatures } from '../../experimental-features';
+import { useExperimentalFeatures } from './experimental-features';
 import menuItemsFixture from './fixtures/menu-items-request-fixture.json';
 
 const TYPE_NAMES = {
