+++ packages/e2e-tests/specs/site-editor/site-editor-export.test.js
@@ -13,7 +13,7 @@ import { trashAllPosts, activateTheme } from '@wordpress/e2e-test-utils';
 /**
  * Internal dependencies
  */
-import { siteEditor } from '../../experimental-features';
+import { siteEditor } from './utils';
 
 async function waitForFileExists( filePath, timeout = 10000 ) {
 	const start = Date.now();
