+++ packages/e2e-tests/specs/site-editor/template-part.test.js
@@ -16,7 +16,7 @@ import {
 /**
  * Internal dependencies
  */
-import { siteEditor } from '../../experimental-features';
+import { siteEditor } from './utils';
 
 const templatePartNameInput =
 	'.edit-site-create-template-part-modal .components-text-control__input';
