+++ packages/e2e-tests/specs/site-editor/template-revert.test.js
@@ -14,7 +14,7 @@ import { addQueryArgs } from '@wordpress/url';
 /**
  * Internal dependencies
  */
-import { siteEditor } from '../../experimental-features';
+import { siteEditor } from './utils';
 
 const {
 	visit: visitSiteEditor,
